Update function for Supplier, Member, Administrator

+++ 3TeamProject/Areas/Suppliers/Controllers/SupplierController.cs
@@ -1,13 +1,10 @@
-﻿using _3TeamProject.Areas.Administrators.Data;
-using _3TeamProject.Areas.Sppliers.Data;
+﻿using _3TeamProject.Areas.Sppliers.Data;
 using _3TeamProject.Areas.Suppliers.Data;
 using _3TeamProject.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 using Microsoft.EntityFrameworkCore;
 using System.Data;
-using System.Linq;
 using System.Net.Mail;
 using System.Security.Claims;
 using System.Security.Cryptography;
@@ -15,7 +12,9 @@ using System.Security.Cryptography;
 
 namespace _3TeamProject.Areas.Sppliers.Controllers
 {
-    [Area("Suppliers")]
+    [Authorize("Suppliers")]
+    [Route("Suppliers/[controller]")]
+    [ApiController]
     public class SupplierController : Controller
     {
         private readonly _3TeamProjectContext _context;
@@ -28,11 +27,15 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
             _config = config;
         }
 
-        [Authorize(Roles = ("Suppliers"))]
-        public IActionResult GetSupplier()
+        [HttpGet("{id}")]
+        public IActionResult GetSupplier(int id)
         {
-            var UserId = User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Sid).Value;
-            var user = _context.Users.Include(u => u.Suppliers).FirstOrDefault(x => x.UserId == int.Parse(UserId));
+            var UserId = int.Parse(User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Sid).Value);
+            if (id != UserId)
+            {
+                return BadRequest("與登入帳號不符");
+            }
+            var user = _context.Users.Include(u => u.Suppliers).FirstOrDefault(x => x.UserId == UserId);
             if (user == null)
             {
                 return BadRequest("沒有此帳號");
@@ -59,33 +62,8 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
                             }).SingleOrDefault();
             return Ok(supplier);
         }
-
-        [Authorize(Roles = ("Administrator, ChiefAdministrator, SuperAdministrator"))]
-        public IActionResult GetAllSuppliers() 
-        {
-                var supplier = from u in _context.Users
-                               join s in _context.Suppliers
-                               on u.UserId equals s.UserId
-                               select new SupplierGetViewModel
-                               {
-                                   Account = u.Account,
-                                   Email = u.Email,
-                                   RoleName = u.RolesNavigation.RoleName,
-                                   ContactName = s.ContactName,
-                                   CompanyName = s.CompanyName,
-                                   TaxId = s.TaxId,
-                                   Fax = s.Fax,
-                                   CellPhoneNumber = s.CellPhoneNumber,
-                                   SupplierPhoneNumber = s.SupplierPhoneNumber,
-                                   SupplierPostalCode = s.SupplierPostalCode,
-                                   SupplierCountry = s.SupplierCountry,
-                                   SupplierCity = s.SupplierCity,
-                                   SupplierAddress = s.SupplierAddress
-                               };
-                return Ok(supplier);
-        }
-
-        public async Task<IActionResult> Register([FromBody]SupplierRequestViewModel request)
+        [HttpPost]
+        public async Task<IActionResult> Register([FromBody] SupplierRegisterViewModel request)
         {
             if (!ModelState.IsValid)
             {
@@ -102,28 +80,6 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
                 var passwordHsah = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(request.Password));
                 var verifyToken = Convert.ToHexString(RandomNumberGenerator.GetBytes(64));
 
-                //Supplier supplier = new Supplier
-                //{
-                //    ContactName = request.ContactName,
-                //    CompanyName = request.CompanyName,
-                //    TaxId = request.TaxId,
-                //    Fax = request.Fax,
-                //    CellPhoneNumber = request.CellPhoneNumber,
-                //    SupplierPhoneNumber = request.SupplierPhoneNumber,
-                //    SupplierPostalCode = request.SupplierPostalCode,
-                //    SupplierCountry = request.SupplierCountry,
-                //    SupplierCity = request.SupplierCity,
-                //    SupplierAddress = request.SupplierAddress,
-                //    IdNavigation = new User
-                //    {
-                //        Account = request.Account,
-                //        Email = request.Email,
-                //        PasswordHash = passwordHsah,
-                //        PasswordSalt = passwordSalt,
-                //        VerficationToken = verifyToken,
-                //        Roles = request.Roles
-                //    }
-                //};
                 Supplier supplier = new Supplier
                 {
                     User = new User
@@ -138,21 +94,21 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
                 };
 
                 #region Send Email with verify code (正式再解開註解)
-                //using (MailMessage mail = new MailMessage())
-                //{
-                //    mail.From = new MailAddress("dotnettgm102@gmail.com", "帳號驗證碼");
-                //    mail.To.Add("dotnettgm102@gmail.com");
-                //    mail.Priority = MailPriority.Normal;
-                //    mail.Subject = "帳號驗證碼";
-                //    mail.Body = $"<a href=\"https://localhost:7190/User/Verify\"  value=\"{verifyToken}\">帳號驗證碼</a>";
-                //    mail.IsBodyHtml = true;
-                //    SmtpClient MySmtp = new SmtpClient("smtp.gmail.com", 587);
-                //    MySmtp.UseDefaultCredentials = false;
-                //    MySmtp.Credentials = new System.Net.NetworkCredential(_config["mail:Account"], _config["mail:Password"]);
-                //    MySmtp.EnableSsl = true;
-                //    MySmtp.Send(mail);
-                //    MySmtp = null;
-                //}; 
+                using (MailMessage mail = new MailMessage())
+                {
+                    mail.From = new MailAddress("dotnettgm102@gmail.com", "帳號驗證碼");
+                    mail.To.Add("dotnettgm102@gmail.com");
+                    mail.Priority = MailPriority.Normal;
+                    mail.Subject = "帳號驗證碼";
+                    mail.Body = $"<a href=\"https://localhost:7190/User/Verify\"  value=\"{verifyToken}\">帳號驗證碼</a>";
+                    mail.IsBodyHtml = true;
+                    SmtpClient MySmtp = new SmtpClient("smtp.gmail.com", 587);
+                    MySmtp.UseDefaultCredentials = false;
+                    MySmtp.Credentials = new System.Net.NetworkCredential(_config["mail:Account"], _config["mail:Password"]);
+                    MySmtp.EnableSsl = true;
+                    MySmtp.Send(mail);
+                    MySmtp = null;
+                };
                 #endregion
 
                 //_context.Suppliers.Add(supplier);
@@ -161,9 +117,14 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
                 return Ok("註冊成功，請等待驗證信件");
             }
         }
-        [Authorize(Roles = ("Suppliers, Administrator, ChiefAdministrator, SuperAdministrator"))]
-        public async Task<IActionResult> Update(int? id, [FromBody]SupplierUpdateViewModel request)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(int id, [FromBody] SupplierUpdateViewModel request)
         {
+            var UserId = int.Parse(User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Sid).Value);
+            if (id != UserId)
+            {
+                return BadRequest("與登入帳號不符");
+            }
             if (!ModelState.IsValid)
             {
                 var errors = ModelState.Select(x => x.Value.Errors).Where(y => y.Count > 0).ToList();
@@ -171,11 +132,7 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
             }
             var supplier = _context.Suppliers.Include(s => s.User).Where(s => s.UserId == id)
                 .Select(s => s).SingleOrDefault();
-            if (supplier == null)
-            {
-                return BadRequest("此帳號不存在");
-            }
-            //var UserId = int.Parse(User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Sid).Value);
+
             using (var hmac = new HMACSHA512())
             {
                 var passwordSalt = hmac.Key;
@@ -198,12 +155,28 @@ namespace _3TeamProject.Areas.Sppliers.Controllers
             }
             return Ok("修改成功!");
         }
-        public async Task<IActionResult> Delete(int id)
+        [HttpDelete]
+        public IActionResult Delete(int id)
         {
+            var UserId = int.Parse(User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Sid).Value);
+            if (id != UserId)
+            {
+                return BadRequest("與登入帳號不符");
+            }
             var user = _context.Users.Include(u => u.Suppliers).FirstOrDefault(x => x.UserId == id);
             _context.Users.Remove(user);
-            await _context.SaveChangesAsync();
+            _context.SaveChanges();
             return Ok("此帳號已刪除");
+        }
+        [HttpGet("{id}")]
+        public IActionResult GetProduct(int id) //TODO 廠商管理商品
+        {
+            return Ok();
+        }
+        [HttpPost]
+        public IActionResult UploadProduct() //TODO 商品上架
+        {
+            return Ok();
         }
     }
 }
